Add encondings file in utils, fix import statements
