+++ naslib/search_spaces/nasbench101/graph.py
@@ -10,7 +10,7 @@ from naslib.search_spaces.core.query_metrics import Metric
 from naslib.search_spaces.nasbench101.conversions import convert_spec_to_model, convert_spec_to_tuple, \
     convert_tuple_to_spec
 from naslib.search_spaces.nasbench101.encodings import encode_101, encode_101_spec
-from naslib.predictors.utils.encodings import EncodingType
+from naslib.utils.encodings import EncodingType
 from naslib.utils import get_dataset_api
 
 INPUT = "input"
