+++ naslib/search_spaces/nasbench301/encodings.py
@@ -1,7 +1,7 @@
 import numpy as np
 import logging
 
-from naslib.predictors.utils.encodings import EncodingType
+from naslib.utils.encodings import EncodingType
 
 """
 These are the encoding methods for DARTS.
