+++ naslib/search_spaces/nasbench301/graph.py
@@ -20,7 +20,7 @@ from naslib.search_spaces.nasbench301.conversions import (
     make_compact_mutable,
     make_compact_immutable,
 )
-from naslib.predictors.utils.encodings import EncodingType
+from naslib.utils.encodings import EncodingType
 from naslib.search_spaces.core.query_metrics import Metric
 from naslib.search_spaces.nasbench301.encodings import encode_darts, encode_darts_compact
 from .primitives import FactorizedReduce
