+++ naslib/search_spaces/nasbenchnlp/graph.py
@@ -13,7 +13,7 @@ from naslib.search_spaces.nasbenchnlp.conversions import convert_recipe_to_compa
 make_compact_mutable, convert_compact_to_recipe
 from naslib.utils.utils import get_project_root
 from naslib.search_spaces.nasbenchnlp.encodings import encode_nlp
-from naslib.predictors.utils.encodings import EncodingType
+from naslib.utils.encodings import EncodingType
 
 
 HIDDEN_TUPLE_SIZE = 2
